+++ react-app/src/components/Reviews/SingleReview.js
@@ -1,8 +1,10 @@
-import React, { useEffect, useState } from 'react'
+import React, { useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
-import { getReviews, deleteReview } from '../../store/reviews';
+import { getReviews } from '../../store/reviews';
 import './ReviewForm.css';
+
+import ReviewShowForm from './ReviewShowForm'
 
 
 function SingleReview() {
@@ -11,94 +13,18 @@ function SingleReview() {
     const dispatch = useDispatch();
     const reviewState = useSelector(state => state.reviews)
     const reviews = Object.values(reviewState);
-    
-
-    const [showForm, setShowForm] = useState(false);
-    const [deleteButton, showDeleteButton] = useState(false);
-    const [rating, setRating] = useState(0);
-    const [content, setContent] = useState('');
 
 
     useEffect(() => {
         dispatch(getReviews())
     }, [dispatch])
 
-    const handleDelete = async (id) => {
-        reviews.map((review) => {
-            if(user.id == review.userId) {
-                showDeleteButton(!deleteButton)
-            }
-    })
-
-        dispatch(deleteReview(id));
-        dispatch(getReviews())
-    }
-
-    const openForm = () => {
-        reviews.map((review) => {
-            if(user.id == review.userId) {
-                setShowForm(!showForm)
-            }
-        })
-    }
-
-    const handleSubmit = async (e) => {
-        e.preventDefault();
-
-        const review = {
-            ...reviews,
-            rating,
-            content,
-            userId: user.id,
-            // productId
-        }
-    }
-
     return (
         <div className='outer-div'>
             <div>
                 {reviews.map((review) =>
-                <div className='reviewsAll' key={review.id}>
-                    <li>ID: {review.id}</li>
-                    <li>Rating: {review.rating}</li>
-                    <li>{review.content}</li>
-                    <li>User ID: {review.userId}</li>
-                    <li>Product ID: {review.productId}</li>
-                    <div>
-                        <button onClick={() => history.push(`/reviews/review_form/${review.productId}`)}>Review Form</button>
-                    </div>
-                    <div>
-                        <button className='open-edit-form' onClick={openForm}>Edit Review</button>
-                    </div>
-                    <div>
-                        <button onClick={() => handleDelete(review.id)}>Delete Review</button>
-                    </div>
-                    {showForm && (
-                    <div className='form-popup' id='myForm'>
-                        <form className='review-edit-form' onSubmit={handleSubmit}>
-                            <h2>Edit Your Review</h2>
-                            <div className="rate">
-                                <input type="radio" id="star5" name="rate" value="5" onChange={(e) => setRating(e.target.value)}/>
-                                <label htmlFor="star5" title="text">5 stars</label>
-                                <input type="radio" id="star4" name="rate" value="4" onChange={(e) => setRating(e.target.value)} />
-                                <label htmlFor="star4" title="text">4 stars</label>
-                                <input type="radio" id="star3" name="rate" value="3" onChange={(e) => setRating(e.target.value)} />
-                                <label htmlFor="star3" title="text">3 stars</label>
-                                <input type="radio" id="star2" name="rate" value="2" onChange={(e) => setRating(e.target.value)} />
-                                <label htmlFor="star2" title="text">2 stars</label>
-                                <input type="radio" id="star1" name="rate" value="1" onChange={(e) => setRating(e.target.value)} />
-                                <label htmlFor="star1" title="text">1 star</label>
-                            </div>
-                            <br></br>
-                            <br></br>
-                            <div className='review-box'>
-                                <textarea placeholder='Your Review' rows='10' cols='50' value={content} onChange={(e) => setContent(e.target.value)}></textarea>
-                            </div>
-                            <button type='submit' className='submit-button'>Submit</button>
-                        </form>
-                    </div>
-                    )}
-                </div>)}
+                    <ReviewShowForm review={review}/>
+                )}
             </div>
         </div>
     )
